Spring : 게시판만들기 회원제 게시판 마무리, 엑셀다운로드

+++ Spring/hello-spring/src/main/java/com/hello/forum/bbs/web/BoardController.java
@@ -1,5 +1,16 @@
 package com.hello.forum.bbs.web;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
+
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.ResponseEntity;
@@ -10,13 +21,17 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.hello.forum.bbs.service.BoardService;
 import com.hello.forum.bbs.vo.BoardListVO;
 import com.hello.forum.bbs.vo.BoardVO;
 import com.hello.forum.beans.FileHandler;
+import com.hello.forum.member.vo.MemberVO;
 import com.hello.forum.utils.ValidationUtils;
+
+import jakarta.servlet.http.HttpSession;
 
 //import jakarta.validation.Valid;
 
@@ -51,7 +66,12 @@ public class BoardController {
 	 * @return
 	 */
 	@GetMapping("/board/write")
-	public String viewBoardWritePage() {
+	public String viewBoardWritePage(HttpSession session) {
+//		MemberVO memberVO = (MemberVO) session.getAttribute("_LOGIN_USER_");
+//		if(memberVO == null) {
+//			return "redirect:/member/login";
+//		}
+		
 		return "board/boardwrite";
 	}
 	
@@ -87,6 +107,7 @@ public class BoardController {
 			// @Valid 에 의해 실행된 파라미터 검사 (NotEmpty, Size, Min, Max, 등)의 결과
 //			BindingResult bindingResult,
 			@RequestParam MultipartFile file,
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO,
 			Model model
 			) {
 		System.out.println("글 등록 처리를 해야합니다.");
@@ -116,9 +137,9 @@ public class BoardController {
 		// 수동 검사 시작
 		// 제목 검사
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
+		
 		boolean isNotEmptycontent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
+		
 		
 		if(!isNotEmptySubject) {
 			//제목을 입력하지 않았다면...
@@ -126,24 +147,14 @@ public class BoardController {
 			model.addAttribute("boardVO",boardVO);
 			return "board/boardwrite";
 		}
-		if(!isNotEmptyEmail) {
-			//이메일을 입력하지 않았다면...
-			model.addAttribute("errorMessage","이메일은 필수 입력 값입니다.");
-			model.addAttribute("boardVO",boardVO);
-			return "board/boardwrite";			
-		} 
 		if(!isNotEmptycontent) {
 			//내용을 입력하지 않았다면...
 			model.addAttribute("errorMessage","내용은 필수 입력 값입니다.");
 			model.addAttribute("boardVO",boardVO);
 			return "board/boardwrite";
 		}
-		if(!isEmailFormat) {
-			//이메일을 이메일 형태로 입력하지 않았다면...
-			model.addAttribute("errorMessage","이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO",boardVO);
-			return "board/boardwrite";			
-		}
+		
+		boardVO.setEmail(memberVO.getEmail());
 		
 		boolean isCreateSuccess = this.boardService.createNewBoard(boardVO, file);
 		if(isCreateSuccess) {
@@ -181,9 +192,13 @@ public class BoardController {
 	}
 	
 	@GetMapping("/board/modify/{id}")  // /board/modify/1 <-- id 변수의 값은 1
-	public String viewBoardModifyPage(@PathVariable int id, Model model) {
+	public String viewBoardModifyPage(@PathVariable int id, @SessionAttribute("_LOGIN_USER_") MemberVO memberVO, Model model) {
 		// 1. 전달받은 id 의 값으로 게시글을 조회한다.
 		BoardVO boardVO = this.boardService.getOneBoard(id, false);
+		
+		if(!memberVO.getEmail().equals(boardVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다!");
+		}
 		
 		// 2. 게시글의 정보를 화면에 보내준다.
 		model.addAttribute("boardVO", boardVO);
@@ -199,14 +214,18 @@ public class BoardController {
 	 * @return 
 	 */
 	@PostMapping("/board/modify/{id}")
-	public String doboardModify(@PathVariable int id, BoardVO boardVO, @RequestParam MultipartFile file, Model model) {
+	public String doboardModify(@PathVariable int id, BoardVO boardVO, @RequestParam MultipartFile file, @SessionAttribute("_LOGIN_USER_") MemberVO memberVO, Model model) {
+		
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		
+		if(!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다!");
+		}
 		
 		// 수동 검사 시작
 		// 제목 검사
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
 		boolean isNotEmptycontent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 		
 		if(!isNotEmptySubject) {
 			//제목을 입력하지 않았다면...
@@ -214,28 +233,28 @@ public class BoardController {
 			model.addAttribute("boardVO",boardVO);
 			return "board/boardmodify";
 		}
-		if(!isNotEmptyEmail) {
-			//이메일을 입력하지 않았다면...
-			model.addAttribute("errorMessage","이메일은 필수 입력 값입니다.");
-			model.addAttribute("boardVO",boardVO);
-			return "board/boardmodify";			
-		} 
+//		if(!isNotEmptyEmail) {
+//			//이메일을 입력하지 않았다면...
+//			model.addAttribute("errorMessage","이메일은 필수 입력 값입니다.");
+//			model.addAttribute("boardVO",boardVO);
+//			return "board/boardmodify";			
+//		} 
 		if(!isNotEmptycontent) {
 			//내용을 입력하지 않았다면...
 			model.addAttribute("errorMessage","내용은 필수 입력 값입니다.");
 			model.addAttribute("boardVO",boardVO);
 			return "board/boardmodify";
 		}
-		if(!isEmailFormat) {
-			//이메일을 이메일 형태로 입력하지 않았다면...
-			model.addAttribute("errorMessage","이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO",boardVO);
-			return "board/boardmodify";			
-		}
+//		if(!isEmailFormat) {
+//			//이메일을 이메일 형태로 입력하지 않았다면...
+//			model.addAttribute("errorMessage","이메일을 올바른 형태로 작성해주세요.");
+//			model.addAttribute("boardVO",boardVO);
+//			return "board/boardmodify";			
+//		}
 		// Command Object 에는 전달된 ID가 없으므로
 		// PathVariable로 전달된 ID를 셋팅해준다.
 		boardVO.setId(id);
-		
+				
 		boolean isUpdatedSuccess= this.boardService.updateOneBoard(boardVO, file);
 		
 		if(isUpdatedSuccess) {
@@ -264,7 +283,13 @@ public class BoardController {
 	 * 
 	 */
 	@GetMapping("/board/delete/{id}")
-	public String doDeleteBoard(@PathVariable int id) {
+	public String doDeleteBoard(@PathVariable int id, @SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
+		
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		
+		if(!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다!");
+		}
 		
 		boolean isDeletedSuccess = this.boardService.deleteOneBoard(id);
 		
@@ -296,5 +321,95 @@ public class BoardController {
 		// 첨부된 파일이 있을 겨웅엔 파일을 사용자에게 보여준다. (Download)
 		
 		return this.fileHandler.download(boardVO.getOriginFileName(), boardVO.getFileName());
+	}
+	
+	@GetMapping("/board/excel/download")
+	public ResponseEntity<Resource> downloadExcelFile() {
+		// 모든 게시글 조회
+		BoardListVO boardListVO = boardService.getAllBoard();
+		
+		// XLSX 문서 만들기
+		Workbook workbook = new SXSSFWorkbook(-1);
+		
+		// 엑셀 시트 만들기
+		Sheet sheet = workbook.createSheet("게시글 목록");
+		
+		// 행 만들기
+		Row row = sheet.createRow(0);
+		
+		// 타이틀 만들기
+		Cell cell = row.createCell(0);
+		cell.setCellValue("번호");
+		
+		cell = row.createCell(1);
+		cell.setCellValue("제목");
+		
+		cell = row.createCell(2);
+		cell.setCellValue("첨부파일명");
+		
+		cell = row.createCell(3);
+		cell.setCellValue("작성자이메일");
+		
+		cell = row.createCell(4);
+		cell.setCellValue("조회수");
+		
+		cell = row.createCell(5);
+		cell.setCellValue("등록일");
+		
+		cell = row.createCell(6);
+		cell.setCellValue("수정일");
+		
+		// 데이터 행 만들고 쓰기
+		List<BoardVO> boardList = boardListVO.getBoardList();
+		int rowIndex = 1;
+		for(BoardVO boardVO : boardList) {
+			row = sheet.createRow(rowIndex);
+			cell = row.createCell(0);
+			cell.setCellValue("" + boardVO.getId());
+			
+			cell = row.createCell(1);
+			cell.setCellValue("" + boardVO.getSubject());
+			
+			cell = row.createCell(2);
+			cell.setCellValue("" + boardVO.getOriginFileName());
+			
+			cell = row.createCell(3);
+			cell.setCellValue("" + boardVO.getEmail());
+			
+			cell = row.createCell(4);
+			cell.setCellValue("" + boardVO.getViewCnt());
+			
+			cell = row.createCell(5);
+			cell.setCellValue("" + boardVO.getCrtDt());
+			
+			cell = row.createCell(6);
+			cell.setCellValue("" + boardVO.getMdfyDt());
+			
+			rowIndex += 1;
+		}
+		
+		// 엑셀 파일 만들기
+		File storedFile = fileHandler.getStoredFile("게시글_목록.xlsx");
+		OutputStream os = null;
+		try {
+			os = new FileOutputStream(storedFile);
+			workbook.write(os);
+		} catch (IOException e) {
+			throw new IllegalArgumentException("엑셀파일을 만들 수 없습니다.");
+		} finally {
+			try {
+				workbook.close();
+			} catch(IOException e) {}
+			if(os != null) {
+				try {
+					os.flush();
+				} catch(IOException e) {}
+				try {
+					os.close();
+				} catch(IOException e) {}
+			}
+		}
+		
+		return this.fileHandler.download("게시글_목록.xlsx", "게시글_목록.xlsx");
 	}
 }
